CPLP-1306 cleanup

+++ src/framework/CatenaX.NetworkServices.Framework.ErrorHandling/GeneralHttpErrorHandler.cs
@@ -73,7 +73,7 @@ public class GeneralHttpErrorHandler
                 statusCode = HttpStatusCode.InternalServerError;
                 logLevel = LogLevel.Error;
             }
-            _logger.Log(logLevel, error, "GeneralErrorHandler caught {error} resulting in response status code {statusCode}, message '{message}'", error.GetType().Name, statusCode, error.Message);
+            _logger.Log(logLevel, error, "GeneralErrorHandler caught {error} resulting in response status code {statusCode}, message '{message}'", error.GetType().Name, (int)statusCode, error.Message);
             context.Response.ContentType = "application/json";
             context.Response.StatusCode = (int)statusCode;
             await context.Response.WriteAsync(JsonSerializer.Serialize(CreateErrorResponse(statusCode, error, messageFunc))).ConfigureAwait(false);
